@@ -8,7 +8,8 @@
     "start:dev": "cross-env NODE_ENV=development webpack-dev-server",
     "start:staging": "cross-env NODE_ENV=staging webpack-dev-server",
     "start:production": "cross-env NODE_ENV=production webpack-dev-server",
-    "test": "echo \"Error: no test specified\" && exit 1"
+    "test": "echo \"Error: no test specified\" && exit 1",
+    "test:lint": "tslint --project ."
   },
   "repository": {
     "type": "git",
@@ -17,43 +18,43 @@
   "author": "csh4017",
   "license": "ISC",
   "dependencies": {
-    "@material-ui/core": "^3.9.2",
-    "@material-ui/icons": "^3.0.2",
-    "@types/react-router": "^4.4.5",
-    "@types/react-router-dom": "^4.3.1",
+    "@material-ui/core": "^4.0.1",
+    "@material-ui/icons": "^4.0.1",
+    "@types/react-router": "^5.0.1",
+    "@types/react-router-dom": "^4.3.3",
     "cross-env": "^5.2.0",
     "jsonwebtoken": "^8.5.1",
     "lodash": "^4.17.11",
-    "mongodb": "^3.1.13",
-    "react": "^16.8.4",
-    "react-dom": "^16.8.4",
-    "react-redux": "^6.0.1",
+    "mongodb": "^3.2.6",
+    "react": "^16.8.6",
+    "react-dom": "^16.8.6",
+    "react-redux": "^7.0.3",
     "react-router": "^5.0.0",
     "react-router-dom": "^5.0.0",
     "redux": "^4.0.1",
     "redux-thunk": "^2.3.0",
-    "typescript": "^3.3.3333"
+    "typescript": "^3.4.5"
   },
   "devDependencies": {
-    "@types/react": "^16.8.8",
-    "@types/react-dom": "^16.8.2",
-    "@types/react-redux": "^7.0.5",
+    "@types/react": "^16.8.18",
+    "@types/react-dom": "^16.8.4",
+    "@types/react-redux": "^7.0.9",
     "awesome-typescript-loader": "^5.2.1",
     "classnames": "^2.2.6",
     "css-loader": "^2.1.1",
     "file-loader": "^3.0.1",
     "html-webpack-plugin": "^3.2.0",
-    "node-sass": "^4.11.0",
+    "node-sass": "^4.12.0",
     "sass-loader": "^7.1.0",
     "source-map-loader": "^0.2.4",
     "style-loader": "^0.23.1",
-    "terser-webpack-plugin": "^1.2.3",
-    "tslint": "^5.14.0",
-    "tslint-react": "^3.6.0",
-    "webpack": "^4.29.6",
-    "webpack-cli": "^3.3.0",
-    "webpack-dev-middleware": "^3.6.1",
-    "webpack-dev-server": "^3.2.1",
-    "webpack-hot-middleware": "^2.24.3"
+    "terser-webpack-plugin": "^1.3.0",
+    "tslint": "^5.16.0",
+    "tslint-react": "^4.0.0",
+    "webpack": "^4.32.2",
+    "webpack-cli": "^3.3.2",
+    "webpack-dev-middleware": "^3.7.0",
+    "webpack-dev-server": "^3.4.1",
+    "webpack-hot-middleware": "^2.25.0"
   }
 }
