@@ -1,17 +1,17 @@
-import { combineReducers } from "redux"
-import jwt = require('jsonwebtoken');
+import jwt = require('jsonwebtoken')
+import { combineReducers } from 'redux'
 
-let initialState = {
+const initialState = {
     loggedIn: false,
-    logginIn: false
+    logginIn: false,
 }
 
-if (localStorage.getItem("userToken")){
+if (localStorage.getItem('userToken')) {
 
-    const decoded = jwt.decode(localStorage.getItem("userToken"), {json: true})
+    const decoded = jwt.decode(localStorage.getItem('userToken'), {json: true})
 
-    if(decoded.exp < Date.now() / 1000){
-        localStorage.removeItem("userToken")
+    if (decoded.exp < Date.now() / 1000) {
+        localStorage.removeItem('userToken')
     } else {
         initialState.loggedIn = true
     }
@@ -23,24 +23,24 @@ const user = (state = initialState, action) => {
         case 'loginRequest':
             return {
                 logginIn: true,
-                loggedIn: false
+                loggedIn: false,
             }
         case 'loginSuccess':
             return {
                 logginIn: false,
-                loggedIn: true
+                loggedIn: true,
             }
         case 'loginFailed':
             return {
                 logginIn: false,
-                loggedIn: false
+                loggedIn: false,
             }
         case 'fetchDetailSuccesss':
             return {
                 ...state,
                 email: action.email,
                 fname: action.fname,
-                sname: action.sname
+                sname: action.sname,
             }
         default:
             return state
