@@ -1,17 +1,18 @@
-import authHeader from "./authHeader";
+import authHeader from './authHeader'
 
 const fetchProtected = (path, headers, body, method, callBack) => {
+  const data = {
+    headers: {
+      ...authHeader(),
+      headers,
+    },
+    method,
+    body,
+  }
 
-    let data = {
-        headers: {
-            ...authHeader(),
-            headers
-        },
-        method: method,
-        body: body
-    }
-
-    fetch(path, data).then(res => res.json()).then(res => callBack(res))
+  fetch(path, data)
+    .then((res) => res.json())
+    .then((res) => callBack(res))
 }
 
 export default fetchProtected
