@@ -1,14 +1,14 @@
 import * as React from 'react'
-import { Banner } from './Banner';
-import { Me } from './me';
+import { Banner } from './Banner'
+import { Me } from './me'
 
 const Home = () => {
-    return (
-        <div>
-            <Banner/>
-            <Me />
-        </div>
-    )
+  return (
+    <div>
+      <Banner />
+      <Me />
+    </div>
+  )
 }
 
 export default Home
