@@ -1,34 +1,34 @@
 import * as React from 'react'
-import { Route, Redirect } from 'react-router-dom'
-import { connect } from 'react-redux';
+import { connect } from 'react-redux'
+import { Redirect, Route } from 'react-router-dom'
 
+const PrivateRoute = ({ component: Component, user, ...rest }) => {
+  const renderComponent = (props) =>
+    user.loggedIn ? (
+      <Component {...props} />
+    ) : (
+      <Redirect
+        to={{
+          pathname: '/login',
+          state: { from: props.location },
+        }}
+      />
+    )
 
-const PrivateRoute = ({ component: Component, user,...rest }) => {
-  return (
-    <Route
-      {...rest}
-      render={props =>
-        user.loggedIn ? (
-          <Component {...props} />
-        ) : (
-          <Redirect
-            to={{
-              pathname: "/login",
-              state: { from: props.location }
-            }}
-          />
-        )
-      }
-    />
-  )
+  return <Route {...rest} render={renderComponent} />
 }
 
 const mapStateToProps = (state, ownProps) => {
   return {
-    user: state.user
+    user: state.user,
   }
 }
 
-const connectedPrivateRoute = connect(mapStateToProps, null, null, {pure : false})(PrivateRoute)
+const connectedPrivateRoute = connect(
+  mapStateToProps,
+  null,
+  null,
+  { pure: false },
+)(PrivateRoute)
 
-export {connectedPrivateRoute as PrivateRoute}
+export { connectedPrivateRoute as PrivateRoute }
