@@ -2,8 +2,8 @@
 import MuiThemeProvider from '@material-ui/core/styles/MuiThemeProvider'
 import * as React from 'react'
 import * as ReactDOM from 'react-dom'
-import { Router, Route, Switch } from 'react-router-dom'
 import { Provider } from 'react-redux'
+import { Route, Router, Switch } from 'react-router-dom'
 import { PrivateRoute } from './components/router/privateRoute'
 import './style.scss'
 
@@ -12,12 +12,12 @@ import store from './redux/store'
 
 /* Component Imports */
 import { createMuiTheme } from '@material-ui/core'
+import Home from './components/frontpage/Home'
+import { Login } from './components/frontpage/Login'
+import Register from './components/frontpage/Register'
+import Thanks from './components/frontpage/Thanks'
 import Hello from './components/Hello'
 import NotFound from './components/notfound/NotFound'
-import {Login} from './components/frontpage/Login';
-import Register from './components/frontpage/Register';
-import Thanks from './components/frontpage/Thanks';
-import Home from './components/frontpage/Home';
 
 /* Passing props to a routed page */
 const HelloPage = () => <Hello compiler='TypeScript' framework='React' />
@@ -25,10 +25,10 @@ const HelloPage = () => <Hello compiler='TypeScript' framework='React' />
 const theme = createMuiTheme({
   typography: {
     fontFamily: 'Noto Sans JP',
-  }
+  },
 })
 
-ReactDOM.render(
+ReactDOM.render((
   <Provider store={store}>
     <MuiThemeProvider theme={theme}>
       <Router history={history}>
@@ -41,6 +41,6 @@ ReactDOM.render(
         </Switch>
       </Router>
     </MuiThemeProvider>
-  </Provider>,
-  document.getElementById('example'),
+  </Provider>
+  ), document.getElementById('example'),
 )
