@@ -1,6 +1,6 @@
+import { combineReducers } from 'redux'
 import user from './userReducer'
-import { combineReducers } from 'redux';
 
 export default combineReducers({
-    user
+    user,
 })
